requirements.txt
PySide6==6.7.2
pydantic==2.8.2
python-dateutil==2.9.0.post0
appdirs==1.4.4
tqdm==4.66.5
pyinstaller==6.10.0
